# Backend/services/KudibaInvoicing/docs/KUDIBA_INVOICING_OPENAPI.yaml
openapi: 3.1.0
info:
  title: KudibaInvoicing - Motor Fiscal e Microservico de Facturacao
  version: 1.0.0
  description: >
    Contrato do motor fiscal do ERP Kudiba, escrito em Rust (Axum + Tokio) com Clean Architecture,
    CQRS e Unit of Work sobre PostgreSQL 16 ACID.

    Este servico e interno a cluster: e alcancavel pelo **API Gateway** (que aplica autenticacao JWT,
    resolucao de tenant, rate limiting e bloqueio de contingencia AGT) ou directamente pelos terminais
    POS (Kudiba Edge) em modo offline-first.

    ### Conformidade Regulatoria
    Conforme o **Decreto Presidencial n.o 71/25 de Angola (AGT)**:
    - Assinatura digital RSA 2048-bit PKCS#1 v1.5 com encadeamento de hash SHA-256.
    - 4 caracteres de controlo extraidos das posicoes 1a, 11a, 21a e 31a da assinatura Base64.
    - Bloqueio pessimista por serie fiscal (`SELECT ... FOR UPDATE`) garantindo sequencialidade estrita
      sem lacunas (*zero gaps*).
    - Imutabilidade das faturas com trigger `trg_protect_invoices` no PostgreSQL.

    ### Documentacao interactiva
    A interface de referencia da API e servida pelo **Scalar** em `/scalar` (ou `/docs`).
    A especificacao bruta fica disponivel em `/api-docs/openapi.yaml`.

    ### Contrato gRPC
    O servico expoe tambem o contrato `kudiba.fiscal.v1` (Tonic) na porta 9091:
    `SignDocument`, `VerifySignature`, `ValidateSeriesSequence` e `TriggerSaftGeneration`.

servers:
  - url: http://localhost:8080
    description: Atraves do API Gateway Kudiba (com middlewares de seguranca)
  - url: http://localhost:9090
    description: KudibaInvoicing directo (Rust - sem middlewares de Gateway)

tags:
  - name: Facturacao & Motor Fiscal
    description: Emissao, consulta e assinatura de documentos fiscais
  - name: Series Fiscais & Validacao
    description: Consistência de sequencia, assinatura directa e verificacao independente
  - name: Observabilidade & Documentacao
    description: Health check e documentacao da API

paths:
  # ===========================================================================
  # EMISSAO E CONSULTA DE DOCUMENTOS FISCAIS
  # ===========================================================================
  /api/v1/invoices:
    post:
      summary: Emitir Fatura / Documento Fiscal com Assinatura Digital AGT
      description: >
        Emite um documento com relevancia fiscal (FT, FR, NC, ND) com bloqueio pessimista
        atomico da serie fiscal (`SELECT ... FOR UPDATE`), numeracao rigorosamente sequencial,
        calculo de IVA, encadeamento de hash SHA-256 e assinatura digital RSA 2048-bit.
      tags:
        - Facturacao & Motor Fiscal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueInvoiceRequest'
            example:
              tenantId: "a0000000-0000-0000-0000-000000000001"
              documentType: "FT"
              seriesCode: "KUD26"
              fiscalYear: 2026
              customerName: "Empresa de Transportes Sonangol EP"
              customerNif: "5412345678"
              currency: "AOA"
              lines:
                - productCode: "PROD-001"
                  description: "Licenciamento Kudiba ERP Cloud Anual"
                  quantity: 1
                  unitPrice: 500000.00
                  taxRate: 14.00
                - productCode: "SERV-002"
                  description: "Consultoria e Parametrizacao Fiscal AGT"
                  quantity: 10
                  unitPrice: 25000.00
                  taxRate: 14.00
      responses:
        '201':
          description: Documento fiscal emitido, assinado digitalmente e gravado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueInvoiceResponse'
              example:
                invoiceId: "9c68dc2d-c1d8-4e5b-9a09-3794af2b549a"
                documentNumber: "FT KUD26/000001"
                sequenceNumber: 1
                netTotal: 750000.00
                taxTotal: 105000.00
                grossTotal: 855000.00
                hashSha256: "5e0af1859bb4a4c170fd3ec53980df5b4623277c59e88a7bab1e18795f5366da"
                signatureRsaBase64: "efOP2LZtH7t5y2sGim3ZwMqfn6+TnPUaX4SxrSmcpa5EM8OyThVwc+KfRnxm0FJURCpZWb7GwQMT6UPGM3ySoPM4k5kOEhJ8tAsNacKkWYZq00W1pnzFovRKjBYEhHVf9nMf55kntiEmD/yya2hJ7lk8BUUFCAF/Z8JEVDQxUg2W4n2E6tTMotB0MZNHLE7fpa1VUGEA3PaRblWYZ2HVg71DFZaeIV3cHLc35s9eQnetinZpG892gYITySpH6ljlwA5j99yskyj4xbFbGV6tLk7ovPDZnZHbXVgg1oJnhHi99hskJTA5gJgbcfHyBduFEXote4SlT6j0gAY0XlnijA=="
                validationChars: "etwU"
                issuedAt: "2026-10-03T00:00:00Z"
                systemEntryDate: "2026-10-03T10:11:11.175384Z"
                isContingency: false
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Conflito de concorrencia na serie fiscal (emissao concorrente na mesma serie)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/fiscal/invoices:
    post:
      summary: Emitir Fatura (rota encaminhada pelo Gateway)
      description: >
        Alias de `POST /api/v1/invoices` mantido para as rotas fiscais do API Gateway
        (`/api/v1/fiscal/*`). Produz exactamente o mesmo documento fiscal.
      tags:
        - Facturacao & Motor Fiscal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueInvoiceRequest'
      responses:
        '201':
          description: Documento fiscal emitido e assinado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueInvoiceResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/invoices/{id}:
    get:
      summary: Consultar Fatura por ID
      description: >
        Recupera os dados completos da fatura emitida, incluindo as linhas de artigo
        e o bloco de assinatura digital.
      tags:
        - Facturacao & Motor Fiscal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador UUID da fatura
          schema:
            type: string
            format: uuid
        - name: tenantId
          in: query
          required: false
          description: Tenant owner da fatura (validacao multi-empresa)
          schema:
            type: string
            format: uuid
            example: "a0000000-0000-0000-0000-000000000001"
      responses:
        '200':
          description: Detalhes completos da fatura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/v1/fiscal/invoices/{id}:
    get:
      summary: Consultar Fatura por ID (rota encaminhada pelo Gateway)
      tags:
        - Facturacao & Motor Fiscal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tenantId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalhes completos da fatura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ===========================================================================
  # CHAVE PUBLICA E OPERACOES CRIPTOGRAFICAS
  # ===========================================================================
  /api/v1/public-key:
    get:
      summary: Obter Chave Publica RSA da AGT
      description: >
        Retorna a chave publica RSA 2048-bit registada para validacao independente das
        assinaturas digitais dos documentos emitidos.
      tags:
        - Series Fiscais & Validacao
      responses:
        '200':
          description: Chave publica e metadados de criptografia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicKeyResponse'
              example:
                keyVersion: "1"
                publicKeyPem: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A...\n-----END PUBLIC KEY-----\n"
                algorithm: "RSA-2048 / SHA-256 PKCS#1 v1.5"

  /api/v1/fiscal/public-key:
    get:
      summary: Obter Chave Publica RSA (rota encaminhada pelo Gateway)
      tags:
        - Series Fiscais & Validacao
      responses:
        '200':
          description: Chave publica e metadados de criptografia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicKeyResponse'

  /api/v1/fiscal/series:
    post:
      summary: Abrir serie fiscal
      description: >
        Abre (ou recupera, de forma idempotente) a serie fiscal de um tenant, tipo de documento e
        ano. A emissao de documentos cria a serie automaticamente quando ainda nao existe.
      tags:
        - Series Fiscais & Validacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFiscalSeriesRequest'
            example:
              tenantId: "9f8c1a54-0d2e-4f6b-9a71-2c5d8e3f1b04"
              documentType: "FT"
              seriesCode: "KUD26"
              fiscalYear: 2026
      responses:
        '201':
          description: Serie fiscal aberta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalSeriesOpenedResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /api/v1/fiscal/sign:
    post:
      summary: Assinatura directa de documento em modo contingencia
      description: >
        Reconstroi e assina o buffer canonico AGT de um documento ja emitido por um terminal POS
        em modo contingencia fiscal, sem qualquer escrita na base de dados (equivalente ao RPC
        gRPC `SignDocument`).
      tags:
        - Series Fiscais & Validacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignDirectRequest'
            example:
              documentNumber: "FT POS01/000042"
              invoiceDate: "2026-10-03"
              systemEntryDate: "2026-10-03T10:11:11"
              grossTotal: 855000.00
              previousHash: "5e0af1859bb4a4c170fd3ec53980df5b4623277c59e88a7bab1e18795f5366da"
              keyVersion: "1"
      responses:
        '200':
          description: Documento assinado com o bloco de assinatura fiscal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignDirectResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'

  /api/v1/fiscal/verify-signature:
    post:
      summary: Verificar assinatura de documento fiscal
      description: >
        Valida de forma independente a assinatura RSA de um documento contra o buffer canonico
        reconstruido e devolve os 4 caracteres de controlo (equivalente ao RPC gRPC
        `VerifySignature`).
      tags:
        - Series Fiscais & Validacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifySignatureRequest'
      responses:
        '200':
          description: Resultado da verificacao da assinatura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifySignatureResponse'

  /api/v1/fiscal/validate-series:
    post:
      summary: Validar consistencia sequencial de uma serie fiscal
      description: >
        Confirma se a proxima sequencia e continua e nao contem lacunas (*gaps*),
        devolvendo a ultima sequencia registada e o ultimo hash encadeado
        (equivalente ao RPC gRPC `ValidateSeriesSequence`).
      tags:
        - Series Fiscais & Validacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateSeriesSequenceRequest'
            example:
              tenantId: "a0000000-0000-0000-0000-000000000001"
              seriesCode: "KUD26"
              expectedSequence: 1
              documentYear: 2026
              documentType: "FT"
      responses:
        '200':
          description: Resultado da validacao de sequencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateSeriesSequenceResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # OBSERVABILIDADE & DOCUMENTACAO
  # ===========================================================================
  /health:
    get:
      summary: Verificacao de Vida (Liveness Probe)
      tags:
        - Observabilidade & Documentacao
      responses:
        '200':
          description: Motor fiscal operacional
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "healthy"
                service: "KudibaInvoicing"
                version: "1.0.0"
                framework: "Rust"
                regulation: "Decreto Presidencial n.º 71/25 (AGT Angola)"
                timestamp: "2026-10-03T12:00:00Z"

  /scalar:
    get:
      summary: Scalar API Reference (documentacao interactiva)
      description: >
        Referencia interactiva da API substituindo o antigo Swagger UI. Tambem disponible
        em `/docs`.
      tags:
        - Observabilidade & Documentacao
      responses:
        '200':
          description: Pagina HTML do Scalar API Reference
          content:
            text/html:
              schema:
                type: string

  /api-docs/openapi.yaml:
    get:
      summary: Especificacao OpenAPI 3.1 Bruta
      tags:
        - Observabilidade & Documentacao
      responses:
        '200':
          description: Ficheiro YAML da especificacao OpenAPI
          content:
            application/yaml:
              schema:
                type: string

components:
  schemas:
    IssueInvoiceRequest:
      type: object
      required:
        - tenantId
        - documentType
        - seriesCode
        - fiscalYear
        - customerName
        - lines
      properties:
        tenantId:
          type: string
          format: uuid
          description: Identificador unico da empresa/organizacao emissora
        documentType:
          type: string
          enum: [FT, FR, FP, NC, ND, RC, GR, GT]
          description: Tipo de documento fiscal nos termos da AGT
          example: "FT"
        seriesCode:
          type: string
          description: Codigo identificador da serie documental (ex. KUD26, A, POS01)
          example: "KUD26"
        fiscalYear:
          type: integer
          description: Ano fiscal de emissao
          example: 2026
        customerName:
          type: string
          description: Nome ou razao social do adquirente
          example: "Empresa de Transportes Sonangol EP"
        customerNif:
          type: string
          description: NIF do adquirente ou 'Consumidor Final'
          default: "Consumidor Final"
          example: "5412345678"
        currency:
          type: string
          default: "AOA"
          example: "AOA"
        isContingency:
          type: boolean
          default: false
        keyVersion:
          type: string
          default: "1"
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceLineInput'

    InvoiceLineInput:
      type: object
      required:
        - productCode
        - description
        - quantity
        - unitPrice
        - taxRate
      properties:
        productCode:
          type: string
          example: "PROD-001"
        description:
          type: string
          example: "Licenciamento Kudiba ERP Cloud Anual"
        quantity:
          type: number
          format: double
          example: 1.0
        unitPrice:
          type: number
          format: double
          example: 500000.00
        taxRate:
          type: number
          format: double
          description: Taxa de IVA aplicada em percentagem (ex. 14.00, 5.00, 0.00)
          example: 14.00
        taxExemptionCode:
          type: string
          nullable: true
          description: Motivo legal de isencao AGT (M00 a M99) obrigatorio se a taxa de IVA for 0%
          example: null

    IssueInvoiceResponse:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
        documentNumber:
          type: string
          example: "FT KUD26/000001"
        sequenceNumber:
          type: integer
          example: 1
        netTotal:
          type: number
          format: double
          example: 750000.00
        taxTotal:
          type: number
          format: double
          example: 105000.00
        grossTotal:
          type: number
          format: double
          example: 855000.00
        hashSha256:
          type: string
          description: Hash SHA-256 gerado pelo encadeamento criptografico do Decreto n.º 71/25
        signatureRsaBase64:
          type: string
          description: Assinatura digital RSA 2048-bit em Base64
        validationChars:
          type: string
          description: Os 4 caracteres de validacao impressos no documento (posicoes 1, 11, 21, 31)
          example: "etwU"
        issuedAt:
          type: string
          format: date-time
        systemEntryDate:
          type: string
          format: date-time
        isContingency:
          type: boolean

    InvoiceDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        seriesId:
          type: string
          format: uuid
        documentNumber:
          type: string
        sequenceNumber:
          type: integer
        documentType:
          type: string
        customerName:
          type: string
        customerNif:
          type: string
        currency:
          type: string
        netTotal:
          type: number
        taxTotal:
          type: number
        grossTotal:
          type: number
        hashSha256:
          type: string
        signatureRsaBase64:
          type: string
        validationChars:
          type: string
        keyVersion:
          type: string
        isContingency:
          type: boolean
        issuedAt:
          type: string
          format: date-time
        systemEntryDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineDetail'

    InvoiceLineDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceId:
          type: string
          format: uuid
        lineNumber:
          type: integer
        productCode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        taxRate:
          type: number
        taxExemptionCode:
          type: string
          nullable: true
        lineTotal:
          type: number

    CreateFiscalSeriesRequest:
      type: object
      required:
        - tenantId
        - documentType
        - seriesCode
        - fiscalYear
      properties:
        tenantId:
          type: string
          format: uuid
        documentType:
          type: string
          description: Sigla do tipo de documento (FT, FS, FR, NC, ND, RG, RA)
          example: "FT"
        seriesCode:
          type: string
          example: "KUD26"
        fiscalYear:
          type: integer
          format: int32
          example: 2026

    FiscalSeriesOpenedResponse:
      type: object
      properties:
        seriesId:
          type: string
          format: uuid

    SignDirectRequest:
      type: object
      required:
        - documentNumber
        - invoiceDate
        - systemEntryDate
        - grossTotal
      properties:
        documentNumber:
          type: string
          example: "FT POS01/000042"
        invoiceDate:
          type: string
          description: Data de emissao (YYYY-MM-DD ou RFC 3339)
          example: "2026-10-03"
        systemEntryDate:
          type: string
          description: Instante de registo no sistema (YYYY-MM-DDTHH:MM:SS ou RFC 3339)
          example: "2026-10-03T10:11:11"
        grossTotal:
          type: number
          format: double
          example: 855000.00
        previousHash:
          type: string
          description: Hash do documento imediatamente anterior (vazio para o 1.o da serie)
        keyVersion:
          type: string
          default: "1"

    SignDirectResponse:
      type: object
      properties:
        documentNumber:
          type: string
        signatureBase64:
          type: string
        hashSha256:
          type: string
        validationChars:
          type: string
        signedAt:
          type: string
          format: date-time
        isContingency:
          type: boolean

    VerifySignatureRequest:
      type: object
      required:
        - documentNumber
        - signatureBase64
        - invoiceDate
        - systemEntryDate
        - grossTotal
      properties:
        documentNumber:
          type: string
        signatureBase64:
          type: string
          description: Assinatura digital completa em Base64
        invoiceDate:
          type: string
        systemEntryDate:
          type: string
        grossTotal:
          type: number
          format: double
        previousHash:
          type: string
        keyVersion:
          type: string
          default: "1"

    VerifySignatureResponse:
      type: object
      properties:
        isValid:
          type: boolean
        validationChars:
          type: string
        errorMessage:
          type: string

    ValidateSeriesSequenceRequest:
      type: object
      required:
        - tenantId
        - seriesCode
        - expectedSequence
        - documentYear
      properties:
        tenantId:
          type: string
          format: uuid
        seriesCode:
          type: string
          example: "KUD26"
        expectedSequence:
          type: integer
          format: int64
          description: Sequencia esperada (N + 1)
          example: 1
        documentYear:
          type: integer
          example: 2026
        documentType:
          type: string
          default: "FT"
          description: Tipo de documento fiscal da serie (por omissao FT)

    ValidateSeriesSequenceResponse:
      type: object
      properties:
        isValid:
          type: boolean
        lastRecordedSequence:
          type: integer
          format: int64
        lastDocumentHash:
          type: string
        isGapDetected:
          type: boolean

    PublicKeyResponse:
      type: object
      properties:
        keyVersion:
          type: string
          example: "1"
        publicKeyPem:
          type: string
          example: "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8A...\n-----END PUBLIC KEY-----\n"
        algorithm:
          type: string
          example: "RSA-2048 / SHA-256 PKCS#1 v1.5"

    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          example: "healthy"
        service:
          type: string
          example: "KudibaInvoicing"
        version:
          type: string
        framework:
          type: string
          example: "Rust"
        regulation:
          type: string
        timestamp:
          type: string
          format: date-time

    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          example: "FISCAL_ENGINE_ERROR"
        timestamp:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invoice must have at least one line item."

  responses:
    BadRequestError:
      description: Requisicao invalida ou parametros ausentes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFoundError:
      description: Documento fiscal nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalError:
      description: Falha na emissao fiscal
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'